--- src/screens/signup/modules/signup-content/signup-id-input.module.tsx
@@ -1,25 +1,0 @@
-import { useFormContext, useController } from 'react-hook-form';
-import { useRecoilValue } from 'recoil';
-
-import { SignUpForm } from '../../hooks';
-import { $signUpState, ESignUpStep } from '../../signup.state';
-
-import { SignUpContentInputComponent } from './components';
-
-export const SignUpIdInputModule = () => {
-  const { screenStep } = useRecoilValue($signUpState);
-
-  const { control } = useFormContext<SignUpForm>();
-  const { field, fieldState } = useController({
-    control,
-    name: ESignUpStep.ID_INPUT,
-  });
-
-  return (
-    <SignUpContentInputComponent
-      screenStep={screenStep}
-      field={field}
-      fieldState={fieldState}
-    />
-  );
-};
